LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/history-traversal/persisted-user-state-restoration/scroll-restoration-fragment-scrolling-samedoc-expected.txt


FAIL Manual scroll restoration should take precedent over scrolling to fragment in cross doc navigation assert_equals: should not scroll to fragment expected 0 but got 800
